refactor: move Claude Code executable path detection to adapter layer
- Move getClaudeCodeExecutablePath function from lib/claude.ts to AnthropicClaudeService
- Enhance AnthropicClaudeService constructor to auto-detect executable path when not provided
- Consolidate type guard functions and parseSDKMessage in domain layer
- Remove src/lib/claude.ts as it contained adapter-specific logic
- Update imports to use domain layer functions directly
- Simplify context initialization by removing manual path detection

This improves architectural separation between utility functions (lib) and adapter-specific logic.
