+++ src/core/domain/claude/types.ts
@@ -103,3 +103,61 @@ export function isToolResult(obj: unknown): obj is ToolResult {
     typeof (obj as ToolResult).tool_use_id === "string"
   );
 }
+
+// Helper function to safely parse SDKMessage
+export function parseSDKMessage(data: unknown): SDKMessage | null {
+  if (typeof data !== "object" || data === null) {
+    return null;
+  }
+
+  const obj = data as Record<string, unknown>;
+
+  if (typeof obj.type !== "string") {
+    return null;
+  }
+
+  // Basic validation for different message types
+  switch (obj.type) {
+    case "assistant":
+      if (
+        obj.message &&
+        typeof obj.message === "object" &&
+        obj.message !== null &&
+        typeof obj.session_id === "string"
+      ) {
+        return obj as AssistantMessage;
+      }
+      break;
+    case "user":
+      if (
+        obj.message &&
+        typeof obj.message === "object" &&
+        obj.message !== null &&
+        typeof obj.session_id === "string"
+      ) {
+        return obj as UserMessage;
+      }
+      break;
+    case "result":
+      if (
+        typeof obj.session_id === "string" &&
+        typeof obj.subtype === "string" &&
+        typeof obj.duration_ms === "number" &&
+        typeof obj.is_error === "boolean"
+      ) {
+        return obj as ResultMessage;
+      }
+      break;
+    case "system":
+      if (
+        typeof obj.session_id === "string" &&
+        typeof obj.subtype === "string" &&
+        typeof obj.cwd === "string"
+      ) {
+        return obj as SystemMessage;
+      }
+      break;
+  }
+
+  return null;
+}
